Spam_classifier-V1/requirements.txt
django>=4.2
djangorestframework>=3.14
mongoengine>=0.27
dnspython>=2.4
scikit-learn>=1.3
nltk>=3.8
joblib>=1.3
numpy>=1.24
pandas>=2.0
python-dotenv>=1.0
gunicorn>=21.2